Moved test to to pkg/registry/core/next/ using package 'next_test'
